chore(frontend): Modify and unify Intervention model to fit new API
