+++ frontend/src/app/modules/intervention-form/form/form.component.html
@@ -6,11 +6,11 @@
       <input type="text" hidden formControlName="id" />
 
       <mat-form-field *ngIf="inPrivateMode && intervention">
-        <input type="text" formControlName="date" matInput placeholder="Data utworzenia" />
+        <input type="text" formControlName="creationDate" matInput placeholder="Data utworzenia" />
       </mat-form-field>
 
       <mat-form-field>
-        <input type="text" formControlName="name" matInput placeholder="Imię i nazwisko" />
+        <input type="text" formControlName="fullName" matInput placeholder="Imię i nazwisko" />
       </mat-form-field>
 
       <mat-form-field>
@@ -23,14 +23,14 @@
       <mat-form-field>
         <input
           type="text"
-          formControlName="phone"
+          formControlName="phoneNumber"
           matInput
           placeholder="Numer telefonu"
           pattern="\+?(\d|\s)*"
           required
         />
         <mat-error>
-          <span *ngIf="isControlInvalid('phone')">Numer telefonu jest wymagany</span>
+          <span *ngIf="isControlInvalid('phoneNumber')">Numer telefonu jest wymagany</span>
         </mat-error>
       </mat-form-field>
 
@@ -47,28 +47,28 @@
         </mat-select>
       </mat-form-field>
 
-      <div formGroupName="address" class="address">
+      <section class="address">
         <mat-form-field class="address--street">
           <input type="text" formControlName="street" matInput placeholder="Ulica" required />
           <mat-error>
-            <span *ngIf="isControlInvalid('address.street')">Ulica jest wymagana</span>
+            <span *ngIf="isControlInvalid('street')">Ulica jest wymagana</span>
           </mat-error>
         </mat-form-field>
 
         <mat-form-field class="address--number">
-          <input type="text" formControlName="number" matInput placeholder="Numer lokalu" required />
+          <input type="text" formControlName="streetNumber" matInput placeholder="Numer lokalu" required />
           <mat-error>
-            <span *ngIf="isControlInvalid('address.street')">Numer lokalu jest wymagany</span>
+            <span *ngIf="isControlInvalid('streetNumber')">Numer lokalu jest wymagany</span>
           </mat-error>
         </mat-form-field>
 
         <mat-form-field class="address--city">
           <input type="text" formControlName="city" matInput placeholder="Miejscowość" required />
           <mat-error>
-            <span *ngIf="isControlInvalid('address.street')">Miejscowość jest wymagana</span>
+            <span *ngIf="isControlInvalid('city')">Miejscowość jest wymagana</span>
           </mat-error>
         </mat-form-field>
-      </div>
+      </section>
 
       <button type="submit" color="primary" mat-raised-button>{{ buttonText }}</button>
     </form>
