+++ frontend/src/app/modules/intervention-form/service/intervention-form.service.ts
@@ -3,7 +3,7 @@ import { HttpClient } from '@angular/common/http';
 
 import { Observable } from 'rxjs';
 
-import { InterventionFormData, InterventionPostData } from '@interventionForm/types';
+import { InterventionFormData } from '@shared/domain/intervention.model';
 import { environment } from '@environment';
 
 @Injectable()
@@ -13,6 +13,6 @@ export class InterventionFormService {
   constructor(private http: HttpClient) {}
 
   post(formData: InterventionFormData): Observable<any> {
-    return this.http.post(this.interventionsUrl, new InterventionPostData(formData));
+    return this.http.post(this.interventionsUrl, formData);
   }
 }
