+++ frontend/src/app/modules/interventions/details/details.component.html
@@ -1,48 +1,37 @@
-<mat-card *ngIf="intervention"
-          [withGtmContext]="interventionDetailsGtmContext"
-          [ngClass]="{'mat-elevation-z4': !embedded, 'embedded': !!embedded}">
+<mat-card
+  *ngIf="intervention"
+  [withGtmContext]="interventionDetailsGtmContext"
+  [ngClass]="{ 'mat-elevation-z4': !embedded, embedded: !!embedded }"
+>
+  <div class="title-wrapper">
+    <mat-card-title>Zgłoszenie z dnia {{ intervention.creationDate | date: 'medium' }}</mat-card-title>
+    <mat-card-subtitle>Identyfikator zgłoszenia: {{ intervention.id }}</mat-card-subtitle>
+  </div>
 
-    <div class="title-wrapper">
-        <mat-card-title>Zgłoszenie z dnia {{intervention.creationDate | date: 'medium' }}</mat-card-title>
-        <mat-card-subtitle>Identyfikator zgłoszenia: {{intervention.id}}</mat-card-subtitle>
-    </div>
+  <mat-card-actions align="end">
+    <button mat-raised-button color="primary" title="Edytuj" (click)="navigateToEditView()">
+      {{ embedded ? '' : 'Edytuj' }}<mat-icon>edit</mat-icon>
+    </button>
+    <button mat-raised-button color="warn" title="Usuń" (click)="showDeleteDialog()">
+      {{ embedded ? '' : 'Usuń' }}<mat-icon>delete</mat-icon>
+    </button>
+  </mat-card-actions>
 
-    <mat-card-actions align="end">
-        <button mat-raised-button color="primary" title='Edytuj' (click)="navigateToEditView()">
-            {{embedded ? '' : 'Edytuj'}}<mat-icon>edit</mat-icon>
-        </button>
-        <button mat-raised-button color="warn" title='Usuń' (click)="showDeleteDialog()">
-            {{embedded ? '' : 'Usuń'}}<mat-icon>delete</mat-icon>
-        </button>
-    </mat-card-actions>
+  <mat-list>
+    <mat-list-item> <mat-icon>done</mat-icon> {{ intervention.status }} </mat-list-item>
+    <mat-list-item> <mat-icon>account_circle</mat-icon> {{ intervention.fullName }} </mat-list-item>
+    <mat-list-item> <mat-icon>phone</mat-icon> {{ intervention.phoneNumber }} </mat-list-item>
+    <mat-list-item> <mat-icon>email</mat-icon> {{ intervention.email }} </mat-list-item>
+    <mat-list-item> <mat-icon>location_on</mat-icon> {{ intervention.address }} </mat-list-item>
+  </mat-list>
 
-    <mat-list>
-        <mat-list-item>
-            <mat-icon>done</mat-icon> {{intervention.status}}
-        </mat-list-item>
-        <mat-list-item>
-            <mat-icon>account_circle</mat-icon> {{intervention.fullName}}
-        </mat-list-item>
-        <mat-list-item>
-            <mat-icon>phone</mat-icon> {{intervention.phone}}
-        </mat-list-item>
-        <mat-list-item>
-            <mat-icon>email</mat-icon> {{intervention.email}}
-        </mat-list-item>
-        <mat-list-item>
-            <mat-icon>location_on</mat-icon> {{intervention.address}}
-        </mat-list-item>
-    </mat-list>
+  <mat-card-content>
+    {{ intervention.description }}
 
-    <mat-card-content>
-        {{intervention.description}}
-
-        <section *ngIf='!embedded' class='map-container'>
-            <agm-map [latitude]="intervention.geoLat" [longitude]="intervention.geoLng">
-                <agm-marker [latitude]="intervention.geoLat" [longitude]="intervention.geoLng">
-                </agm-marker>
-            </agm-map>
-        </section>
-    </mat-card-content>
-
+    <section *ngIf="!embedded" class="map-container">
+      <agm-map [latitude]="intervention.geoLat" [longitude]="intervention.geoLng">
+        <agm-marker [latitude]="intervention.geoLat" [longitude]="intervention.geoLng"> </agm-marker>
+      </agm-map>
+    </section>
+  </mat-card-content>
 </mat-card>
