+++ frontend/src/app/modules/interventions/details/details.component.ts
@@ -1,6 +1,6 @@
 import { Component, Inject, Input } from '@angular/core';
 import { Router } from '@angular/router';
-import { Intervention, InterventionRouterState } from '../types';
+import { Intervention, InterventionRouterState } from '@shared/domain/intervention.model';
 import { MatDialog } from '@angular/material/dialog';
 import { DeleteDialog } from './delete.dialog';
 import { InterventionsService } from '../interventions.service';
