+++ frontend/src/app/modules/interventions/edit-form/edit-form.component.ts
@@ -2,12 +2,10 @@ import { Component } from '@angular/core';
 import { formatDate } from '@angular/common';
 import { ActivatedRoute } from '@angular/router';
 
+import { Observable } from 'rxjs';
 import { first, map, tap } from 'rxjs/operators';
 
-import { InterventionFormData } from '@interventionForm/types';
-
-import { Intervention } from '../types';
-import { Observable } from 'rxjs';
+import { Intervention } from '@shared/domain/intervention.model';
 import { InterventionFormContainer } from '@interventionForm/form-container/form-container-base';
 import { InterventionFormService } from '@interventionForm/service/intervention-form.service';
 import { SnackBarManager } from '@shared/services/snack-bar-manager';
@@ -20,16 +18,14 @@ import { InterventionsService } from '../interventions.service';
   providers: [InterventionFormService],
 })
 export class EditFormComponent extends InterventionFormContainer {
-  intervention: InterventionFormData;
-  loading$: Observable<InterventionFormData> = this.interventionsService
-    .getIntervention(this.activatedRoute.params)
-    .pipe(
-      map(toFormData),
-      tap(formData => {
-        this.intervention = formData;
-      }),
-      first(),
-    );
+  intervention: Intervention;
+  loading$: Observable<Intervention> = this.interventionsService.getIntervention(this.activatedRoute.params).pipe(
+    map(toFormData),
+    tap(formData => {
+      this.intervention = formData;
+    }),
+    first(),
+  );
 
   constructor(
     private activatedRoute: ActivatedRoute,
@@ -41,25 +37,10 @@ export class EditFormComponent extends InterventionFormContainer {
   }
 }
 
-const toFormData = (intervention: Intervention | null): InterventionFormData =>
+const toFormData = (intervention?: Intervention): Intervention =>
   intervention
-    ? ({
-        id: intervention.id,
-        date: formatDate(intervention.creationDate, 'medium', 'pl'),
-        name: intervention.fullName,
-        description: intervention.description,
-        phone: intervention.phone,
-        email: intervention.email,
-        status: intervention.status,
-        address: createAddress(intervention.address),
-      } as InterventionFormData)
+    ? {
+        ...intervention,
+        creationDate: formatDate(intervention.creationDate, 'medium', 'pl'),
+      }
     : null;
-
-const createAddress = (address: string) => {
-  const addressParts = address.split(',').map(part => part.trim());
-  return {
-    street: addressParts[0] || '',
-    number: addressParts[1] || '',
-    city: addressParts[2] || '',
-  };
-};
