+++ frontend/src/app/modules/interventions/list/table/table.component.html
@@ -21,7 +21,7 @@
     <!-- Phone Column -->
     <ng-container matColumnDef="phone">
       <mat-header-cell *matHeaderCellDef>Telefon</mat-header-cell>
-      <mat-cell *matCellDef="let element" data-label="Telefon">{{ element.phone }}</mat-cell>
+      <mat-cell *matCellDef="let element" data-label="Telefon">{{ element.phoneNumber }}</mat-cell>
     </ng-container>
 
     <!-- Date Column -->
