+++ frontend/src/app/modules/interventions/list/table/table.component.ts
@@ -2,7 +2,7 @@ import { Component, Input, ViewChild, OnChanges, AfterViewInit } from '@angular/
 import { Router } from '@angular/router';
 import { MatTable, MatTableDataSource, MatPaginator } from '@angular/material';
 
-import { Intervention, InterventionRouterState } from '../../types';
+import { Intervention, InterventionRouterState } from '@shared/domain/intervention.model';
 
 @Component({
   selector: 'app-interventions-table',
